refactor: move modules type to helper
