--- apps/server/src/lib/cronTask/index.ts
@@ -1,1 +1,0 @@
-// there is no clean task here yet.
