--- apps/server/src/modules/tool/type/api.ts
@@ -1,8 +1,0 @@
-// Re-export API schemas and types from schemas module
-export {
-  ToolDetailSchema,
-  ToolSimpleSchema,
-  ToolTagListSchema,
-  type ToolDetailType,
-  type ToolSimpleType
-} from '../schemas/common';
