--- apps/server/src/modules/tool/type/fastgpt.ts
@@ -1,18 +1,0 @@
-// Re-export from validates
-export {
-  FlowNodeInputTypeEnum,
-  WorkflowIOValueTypeEnum,
-  LLMModelTypeEnum,
-  FlowNodeOutputTypeEnum,
-  InputConfigSchema,
-  InputSchema,
-  OutputSchema,
-  type InputConfigType,
-  type InputType,
-  type OutputType
-} from '@/validates/tool';
-
-// Module-specific enum
-export enum SystemInputKeyEnum {
-  systemInputConfig = 'system_input_config'
-}
