--- apps/server/src/modules/tool/type/formatter.ts
@@ -1,9 +1,0 @@
-import type { ToolSetType, ToolType } from '.';
-import { ToolDetailSchema, type ToolDetailType } from './api';
-
-// export const formatToolDetail = (mod: ToolType | ToolSetType): ToolDetailItemType => {
-//   return {
-//     description: mod.description,
-
-//   }
-// };
