+++ packages/helpers/package.json
@@ -1,0 +1,3 @@
+{
+  "name": "@fastgpt-plugin/helper"
+}
