+++ packages/helpers/src/catch.ts
@@ -3,6 +3,7 @@
  * @param fn
  * @returns [result, error]
  */
+
 export async function catchError<T>(fn: () => T): Promise<[Awaited<T> | null, unknown]> {
   try {
     return [await fn(), null];
