+++ packages/helpers/src/tools/types/tool.ts
@@ -1,0 +1,6 @@
+import { z } from 'zod';
+import { ToolCallbackReturnSchema } from '../schemas/tool';
+
+// ==================== Module-specific Schemas (with callback) ====================
+
+export type ToolCallbackReturnSchemaType = z.infer<typeof ToolCallbackReturnSchema>;
