--- packages/shared/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@fastgpt-plugin/shared",
-  "private": true
-}
